refactor: rename password validation utility and update tests
