+++ apps/web/src/utils/password-validator/index.ts
